Move algos to its own package :mountain_railway:
