+++ notebooks/algo/algo01_v1.py
