+++ notebooks/algo/algo02_v0.py
